- move server compontents to their own folder - split out capture class into its own module
