+++ pysparkrpc/server/capture.py
@@ -1,0 +1,13 @@
+import sys
+from io import StringIO
+
+class Capture(list):
+    def __enter__(self):
+        self._stdout = sys.stdout
+        sys.stdout = self._stringio = StringIO()
+        return self
+
+    def __exit__(self, *args):
+        self.extend(self._stringio.getvalue().splitlines())
+        del self._stringio
+        sys.stdout = self._stdout
